+++ client/src/components/forms/dynamic-form/index.js
@@ -5,11 +5,13 @@ export {
   DynamicFormToggle,
   DynamicTextArea,
   DynamicPublicKey,
-  DynamicDatePicker,
   DynamicToolTipWrapper,
 } from "./components";
+export { DynamicDatePicker } from "./DynamicDatePicker";
 export { PowerInterfaces } from "./PowerInterfaces";
 export { SubFormOverrideTile } from "./SubFormOverrideTile";
+export { default as DynamicFetchMultiSelect } from "./DynamicFetchMultiSelect";
+export { default as DynamicFetchSelect } from "./DynamicFetchSelect";
 export {
   ClassicDisabledTile,
   NoClassicGatewaysTile,