+++ client/src/lib/state/appid.js
@@ -1,4 +1,4 @@
-const { transpose, splatContains, isNullOrEmptyString } = require("lazy-z");
+const { transpose, splatContains } = require("lazy-z");
 const {
   setUnfoundResourceGroup,
   updateSubChild,